put single review back in review list to fix thumbs
